Fix a rotary encoder num casting bug

+++ src/counter.rs
@@ -14,7 +14,7 @@ impl<PINS> Counter<PINS> {
 
     pub fn poll(&mut self) -> Option<i8> {
         let count = self.qei.count();
-        let diff = count.wrapping_sub(self.last_count) as i32;
+        let diff = count.wrapping_sub(self.last_count) as i16;
 
         if diff.abs() >= 4 {
             self.last_count = count;
